fix: posts - locations, tags, price

+++ app/src/main/java/com/mnassa/screen/posts/need/create/CreateNeedViewModel.kt
@@ -16,5 +16,5 @@ interface CreateNeedViewModel : MnassaViewModel, ChipsAdapter.ChipSearch, PlaceA
 
     suspend fun getTag(tagId: String): TagModel?
     suspend fun getUser(userId: String): ShortAccountModel?
-    fun createPost(need: String, tags: List<TagModel>, images: List<AttachedImage>, placeId: String?, price: Double?, shareOptions: SharingOptionsController.ShareToOptions)
+    fun createPost(need: String, tags: List<TagModel>, images: List<AttachedImage>, placeId: String?, price: Long?, shareOptions: SharingOptionsController.ShareToOptions)
 }
